Configuración final de la parte de usuarios y los datos personales.
