+++ BackDestiCode/Services/Interfaces/IServiceUsuario.cs
@@ -4,8 +4,7 @@ namespace BackDestiCode.Services.Interfaces
 {
     public interface IServiceUsuario
     {
-        Task<List<DatosPersonalesDto>> GetUsuarios();
-        Task<bool> InsertUsuario(DatosPersonalesDto usuario);
-        Task<bool> UpdateUsuario(DatosPersonalesDto usuario);
+        Task<bool> Registrar(UsuariosDto usuario);
+        Task<bool> Validar(string nombreUsuario, string contrasenia);
     }
 }
